phase2/IA_Chatgpt5.2/requirements.txt
numpy>=2.0
scipy>=1.12
Pillow>=10.0
PySide6>=6.6
